% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aca_refs.R
\docType{data}
\name{aca_refs}
\alias{aca_refs}
\title{Data of ACA Ballot Initiatives}
\format{
A tibble
\describe{
\item{Year}{dbl Year price was recorded}
}
}
\source{
\url{}
}
\usage{
aca_refs
}
\description{
ACA Ballot Data
}
\keyword{datasets}
