% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acaballots-package.R
\docType{package}
\name{acaballots-package}
\alias{acaballots}
\alias{acaballots-package}
\title{acaballots: State Data for ACA Ballot Initiatives}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Rohit Swain \email{roswain2002@gmail.com} (\href{https://github.com/rswatriumvir}{ORCID})

}
\keyword{internal}
